Update FC4 to 2019.09.17-6da6ef8
This latest release of FC4 removes the Node renderer, which had been
compiled into a platform-specific binary.

FC4 releases therefore now include only a single, “universal”
distribution package that *should* run on most UNIX-like systems for
which a modern JRE is available, as the program itself is “just” a
JAR file (an überjar) which is *supposed* to be platform-agnostic.

That allows the update-fc4 script and the formula to be simpler.

https://github.com/FundingCircle/fc4-framework/releases/tag/release_2019.09.17-6da6ef8