Stop update script from finding+using pre-releases
I ran the script as it was, and it found a pre-release; that was the
most recent release. So I thought: we need to filter those out. So I
took a look at the docs for releases in the GitHub REST API v3 and found
that there’s a more straightforward way to get the latest release that’s
not a pre-release: they have a resource specifically for the
“latest release”:

> The latest release is the most recent non-prerelease, non-draft
> release, sorted by the created_at attribute. The created_at attribute
> is the date of the commit used for the release, and not the date when
> the release was drafted or published.

https://developer.github.com/v3/repos/releases/#get-the-latest-release

Perfect!