[CI] Fix params of CircleCI executor
I was reviewing #15 when I noticed that when the jobs referenced the
`macos` executor, they were never supplying a value for the executor’s
param `version` — and that this meant that the default value of the
param was being used for every job instance. This meant that even though
in the workflow it _seemed_ like we were defining three different
instances of the `test` job, and that each instance would run on
different version of MacOS, they were in fact always running on the same
version; the one that was the default value of the param.

This is a fairly well-known risk of defining default values for params:
sometimes one may end up inadvertently using that default and not
realizing it. That’s why it’s often safer to omit a default value and
instead require callers to supply a param.

So that’s the core change I’ve made here: I’ve removed the default value
for the param `version` from the executor.

While I was at it, I made some additional tweaks:

* Renamed the executor param `version` to `xcode-version` just to be a
  little clearer and more explicit
* Changed the type of `xcode-version` from a string to an enumeration,
  just to add a little more rigor/safety to its use
* Updated the versions of Xcode that are allowed/used to the latest
  versions I saw listed in
  https://circleci.com/docs/2.0/testing-ios/#supported-xcode-versions
  for each edition of MacOS.
* Replaced the job params named `version` of type string with a param
  named `e` of type `executor`
  * At first I tried changing `version` to an enum, just like with the
    executor param. But that meant the same param was being defined in
    three places, triggering the rule of three. This approach (the
    executor param) means that the version param can be defined (and
    documented) just once, in one place.
  * Docs here: https://circleci.com/docs/2.0/reusing-config/#executor
* Moved the commented-out audit jobs into their own commented-out
  workflow.
  * This just seems a little cleaner and clearer to me.